requirements.txt
pybluez
attrs
pytest
